requirements.txt
streamlit
requests
pandas
geopy
folium
streamlit-folium
python-dotenv